Refactor CustomPopupContainer
