+++ src/components/mapOverlays/LoaderErrorNotifications.tsx
@@ -1,38 +1,16 @@
 import React, { useEffect, useState } from 'react';
 import { useDvkContext } from '../../hooks/dvkContext';
-import { CustomPopup } from './CustomPopup';
 import errorIcon from '../../theme/img/safetyequipment/error_icon.svg';
-import { debounce } from 'lodash';
-import { getMapCanvasWidth, isMobile } from '../../utils/common';
-import dvkMap from '../DvkMap';
-import { ObjectEvent } from 'ol/Object';
 import { IonCol } from '@ionic/react';
 import { useTranslation } from 'react-i18next';
+import CustomPopup, { CustomPopupContainer } from './CustomPopup';
 
 export const LoadErrorNotifications: React.FC = () => {
   const { t } = useTranslation();
   const { state } = useDvkContext();
   const [visible, setVisible] = useState(false);
-  const [backgroundWidth, setBackgroundWidth] = useState<number>(0);
 
   const handlePopupClose = () => setVisible(false);
-
-  const debouncedBackgroundWidthRefresh = React.useRef(
-    debounce(() => {
-      setBackgroundWidth(getMapCanvasWidth());
-    }, 20)
-  ).current;
-
-  /* Use map canvas size as reference for container width to position container relative to side modal */
-  dvkMap.olMap?.on('change:size', (event: ObjectEvent) => {
-    const { target, key, oldValue } = event;
-    const newValue = target.get(key);
-    // Opening side modal triggers map size changes that have undefined or zero values before calculating final size
-    // Ignore changes in height
-    if (newValue !== undefined && newValue[0] && (!oldValue || oldValue[0] !== newValue[0])) {
-      debouncedBackgroundWidthRefresh();
-    }
-  });
 
   useEffect(() => {
     if (Number(state.response[0]) !== 200 && state.response.length !== 0) {
@@ -41,11 +19,7 @@ export const LoadErrorNotifications: React.FC = () => {
   }, [state.response]);
 
   return (
-    // uses same style as marine warning container, since display is identical
-    <div
-      className="marine-warning-container"
-      style={!isMobile() ? { width: backgroundWidth + 'px', left: 'calc(100% - ' + backgroundWidth + 'px)' } : undefined}
-    >
+    <CustomPopupContainer>
       <CustomPopup isOpen={visible} closePopup={handlePopupClose} icon={errorIcon}>
         <IonCol>
           <p>
@@ -56,6 +30,6 @@ export const LoadErrorNotifications: React.FC = () => {
           </p>
         </IonCol>
       </CustomPopup>
-    </div>
+    </CustomPopupContainer>
   );
 };
